Make Dockerfile conform to Docker practices
This should make the image have a lot less layers and much less size in
total when pulled.

Most RUN steps have been merged into one, so as to reduce the number of
layers and intermediate files gathered.

Base image is got from ruby:2.3 to get ruby updates when they come
available.

Apache is removed, and the tracks site is served through Puma via
$ rails server

The apt-get update/install is handled in one single RUN, removing
/var/lib/apt/lists afterwards, so that the apt-get update info does not
remain in the image.

ADD is replaced by COPY as per https://docs.docker.com/engine/reference/builder/#copy

If many COPY could be collapsed into one, that has been made, to further
reduce the image layers.

The source v2.3.0.zip is removed from the repository, as it can be
obtained as part of the build process *in the same step* as it being
removed (thus, not using space on an intermediate image).

The CMD uses the alternative [""] syntax, which does not perform shell
expansion, as it seems not to be needed.